Theme (and code) optimisations
Background colours have been improved in both modes. These are now an off-white and an off-black, compared to the previous monochromatic theme. This works nicely with the simple title bar(s).

Furthermore, a theme in form of a JSON file has been added to capture the colours of the respective modes (light/dark) and the relevant objects in one place, as opposed to constantly declaring them in the original Python file. This has shortened the code needed for the buttons to a great extent. Further shortening will be done through refactoring, specifically for loops, through which redundant code is going to be eliminated.